fix(kmp): use 0.toByte() for ByteVar null-terminator in native readTextFile
CArrayPointer<ByteVar>.set() requires a Byte argument, not Int.
Using plain 0 caused a Kotlin/Native compile error:
  Unresolved reference — none of the candidates is applicable
  because of a receiver type mismatch.

+++ core/src/nativeMain/kotlin/cz/vutbr/fit/interlockSim/util/PlatformIO.kt
@@ -29,7 +29,7 @@ actual fun readTextFile(path: String): String = memScoped {
 
 		val buffer = allocArray<ByteVar>(size + 1)
 		val bytesRead = fread(buffer, 1.convert(), size.convert(), file)
-		buffer[bytesRead.toInt()] = 0
+		buffer[bytesRead.toInt()] = 0.toByte()
 		buffer.toKString()
 	} finally {
 		fclose(file)
